+++ backend/main.py
@@ -1,59 +1,29 @@
-import shutil
+"""
+FastAPI application entry point.
+
+This is a clean, modular entry point that sets up the FastAPI application
+with middleware, routers, and static files.
+"""
+
 import os
-import traceback
-import re
-import json
-import asyncio
-import gc
-import uuid
-from typing import List
-
-# --- 2. Third-party Library Imports ---
 from dotenv import load_dotenv
-from fastapi import FastAPI, File, UploadFile, BackgroundTasks, Depends, HTTPException, Form, Request, Header
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import FileResponse, StreamingResponse, JSONResponse
 from fastapi.staticfiles import StaticFiles
-from sqlalchemy.orm import Session
-from sqlalchemy.sql import func
-from pydantic import BaseModel
-from svix import Webhook
 
-# --- Load Environment Variables FIRST ---
-load_dotenv() # This loads variables from the .env file at the project root
+# Load environment variables first
+load_dotenv()
 
-# --- 3. Internal/Local Application Imports ---
-from . import models, schemas, crud, auth
-from .database import engine, SessionLocal
-from core.config.loader import load_config
-from core.translation.models.gemini import GeminiModel
-from core.translation.models.openrouter import OpenRouterModel
-from core.config.builder import DynamicConfigBuilder
-from core.translation.job import TranslationJob
-from core.translation.engine import TranslationEngine
-from core.utils.file_parser import parse_document
-from core.prompts.manager import PromptManager
-from core.translation.style_analyzer import (
-    extract_sample_text, 
-    analyze_narrative_style_with_api, 
-    parse_style_analysis, 
-    format_style_for_engine,
-    analyze_glossary_with_api,
-    parse_glossary_analysis
-)
+# Import routers
+from .api.v1 import translation, community, admin, webhooks, announcements
 
-# --- Environment Variables & Constants ---
-ADMIN_SECRET_KEY = os.environ.get("ADMIN_SECRET_KEY", "dev-secret-key")
-CLERK_WEBHOOK_SECRET = os.environ.get("CLERK_WEBHOOK_SECRET")
-
-# --- Database Initialization ---
-# Run database migrations (includes table creation)
+# Database initialization
 try:
     from . import migrations
     migrations.run_migrations()
 except Exception as e:
     print(f"❌ Migration error: {e}")
-    # 마이그레이션 실패해도 앱은 계속 실행
+    # Continue running even if migrations fail
 
 # Auto initialization (categories, etc.)
 try:
@@ -61,12 +31,16 @@ try:
     auto_init.run_auto_init()
 except Exception as e:
     print(f"❌ Auto initialization error: {e}")
-    # 초기화 실패해도 앱은 계속 실행
+    # Continue running even if initialization fails
 
-# --- FastAPI App Initialization ---
-app = FastAPI()
+# Create FastAPI app
+app = FastAPI(
+    title="Context-Aware Translation API",
+    description="AI-powered literary translation service with context awareness",
+    version="1.0.0"
+)
 
-# --- Middleware Configuration ---
+# Configure CORS
 app.add_middleware(
     CORSMiddleware,
     allow_origins=[
@@ -83,1304 +57,23 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# --- Static Files Configuration ---
 # Create uploads directory if it doesn't exist
 UPLOAD_DIR = "uploads/images"
 os.makedirs(UPLOAD_DIR, exist_ok=True)
+os.makedirs("uploads", exist_ok=True)
 
 # Mount static files for serving uploaded images
 app.mount("/static", StaticFiles(directory="uploads"), name="static")
 
-# --- Dependency Injection ---
-def get_db():
-    """Dependency to get a database session for a single request."""
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
+# Include routers
+app.include_router(translation.router)
+app.include_router(community.router)
+app.include_router(admin.router)
+app.include_router(webhooks.router)
+app.include_router(announcements.router)
 
-# --- Helper Functions ---
-def get_model_api(api_key: str, model_name: str, config: dict):
-    """Factory function to get the correct model API instance."""
-    if api_key.startswith("sk-or-"):
-        print(f"--- [API] Using OpenRouter model: {model_name} ---")
-        return OpenRouterModel(
-            api_key=api_key,
-            model_name=model_name,
-            enable_soft_retry=config.get('enable_soft_retry', True)
-        )
-    else:
-        print(f"--- [API] Using Gemini model: {model_name} ---")
-        return GeminiModel(
-            api_key=api_key,
-            model_name=model_name,
-            safety_settings=config['safety_settings'],
-            generation_config=config['generation_config'],
-            enable_soft_retry=config.get('enable_soft_retry', True)
-        )
-
-def validate_api_key(api_key: str, model_name: str):
-    """Validates the API key based on its prefix."""
-    if api_key.startswith("sk-or-"):
-        return OpenRouterModel.validate_api_key(api_key, model_name)
-    else:
-        return GeminiModel.validate_api_key(api_key, model_name)
-
-# --- Background Task Definition ---
-def run_translation_in_background(job_id: int, file_path: str, filename: str, api_key: str, model_name: str, style_data: str = None, glossary_data: str = None, segment_size: int = 15000):
-    db = None
-    try:
-        db = SessionLocal()
-        crud.update_job_status(db, job_id, "PROCESSING")
-        print(f"--- [BACKGROUND] Starting translation for Job ID: {job_id}, File: {filename}, Model: {model_name} ---")
-        
-        config = load_config()
-        gemini_api = get_model_api(api_key, model_name, config)
-
-        translation_job = TranslationJob(
-            file_path, 
-            original_filename=filename, 
-            target_segment_size=segment_size
-        )
-        
-        initial_core_style_text = None
-        protagonist_name = "protagonist"
-        if style_data:
-            try:
-                style_dict = json.loads(style_data)
-                print(f"--- [BACKGROUND] Using user-defined core style for Job ID: {job_id}: {style_dict} ---")
-                protagonist_name = style_dict.get('protagonist_name', 'protagonist')
-                initial_core_style_text = format_style_for_engine(style_dict, protagonist_name)
-            except json.JSONDecodeError:
-                print(f"--- [BACKGROUND] WARNING: Could not decode style_data JSON for Job ID: {job_id}. Proceeding with auto-analysis. ---")
-        else:
-            try:
-                sample_text = extract_sample_text(file_path, method="first_segment", count=15000)
-                config = load_config()
-                model_api = get_model_api(api_key, model_name, config)
-                
-                print(f"--- [BACKGROUND] Performing automatic style analysis for Job ID: {job_id} ---")
-                style_report_text = analyze_narrative_style_with_api(sample_text, model_api, filename)
-                parsed_style = parse_style_analysis(style_report_text)
-                protagonist_name = parsed_style.get('protagonist_name', 'protagonist')
-                initial_core_style_text = format_style_for_engine(parsed_style, protagonist_name)
-                print(f"--- [BACKGROUND] Automatic style analysis complete for Job ID: {job_id} ---")
-            except Exception as e:
-                print(f"--- [BACKGROUND] WARNING: Could not perform automatic style analysis for Job ID: {job_id}. Error: {e} ---")
-
-        initial_glossary = None
-        if glossary_data:
-            try:
-                initial_glossary = json.loads(glossary_data)
-                print(f"--- [BACKGROUND] Using user-defined glossary for Job ID: {job_id} ---")
-            except json.JSONDecodeError:
-                print(f"--- [BACKGROUND] WARNING: Could not decode glossary_data JSON for Job ID: {job_id}. ---")
-
-        dyn_config_builder = DynamicConfigBuilder(gemini_api, protagonist_name, initial_glossary=initial_glossary)
-        engine = TranslationEngine(gemini_api, dyn_config_builder, db=db, job_id=job_id, initial_core_style=initial_core_style_text)
-        
-        engine.translate_job(translation_job)
-        crud.update_job_status(db, job_id, "COMPLETED")
-        print(f"--- [BACKGROUND] Translation finished for Job ID: {job_id}, File: {filename} ---")
-    except Exception as e:
-        if db:
-            error_message = f"An unexpected error occurred: {str(e)}"
-            crud.update_job_status(db, job_id, "FAILED", error_message=error_message)
-        print(f"--- [BACKGROUND] An unexpected error occurred for Job ID: {job_id}, File: {filename}. Error: {e} ---")
-        traceback.print_exc()
-    finally:
-        if db:
-            db.close()
-        gc.collect()
-        print(f"--- [BACKGROUND] Job ID: {job_id} finished. DB session closed and GC collected. ---")
-
-def run_validation_in_background(job_id: int, file_path: str, quick_validation: bool, validation_sample_rate: int):
-    """Background task to run validation on a completed translation."""
-    from core.translation.validator import TranslationValidator
-    
-    db = None
-    try:
-        db = SessionLocal()
-        job = crud.get_job(db, job_id=job_id)
-        if not job:
-            print(f"--- [VALIDATION] Job ID {job_id} not found ---")
-            return
-        
-        # Update status to IN_PROGRESS
-        job.validation_status = "IN_PROGRESS"
-        db.commit()
-        print(f"--- [VALIDATION] Starting validation for Job ID: {job_id} ---")
-        
-        # Get the translated file path
-        unique_base = os.path.splitext(os.path.basename(file_path))[0]
-        original_ext = os.path.splitext(job.filename)[1].lower()
-        
-        if original_ext == '.epub':
-            output_ext = '.epub'
-        else:
-            output_ext = '.txt'
-        
-        translated_filename = f"{unique_base}_translated{output_ext}"
-        translated_path = os.path.join("translated_novel", translated_filename)
-        
-        if not os.path.exists(translated_path):
-            raise FileNotFoundError(f"Translated file not found: {translated_path}")
-        
-        # Initialize validator
-        config = load_config()
-        # Get API key from the job's owner (would need to store this or pass it)
-        # For now, we'll use the environment variable as fallback
-        api_key = os.environ.get("GEMINI_API_KEY", "")
-        model_name = "gemini-2.0-flash-exp"
-        
-        model_api = get_model_api(api_key, model_name, config)
-        validator = TranslationValidator(model_api)
-        
-        # Run validation
-        sample_rate = validation_sample_rate / 100.0  # Convert percentage to decimal
-        validation_job = TranslationJob(file_path, original_filename=job.filename)
-        
-        # Load the translated segments from the translated file
-        # Since we know the translated file is a text file with segments separated by newlines
-        with open(translated_path, 'r', encoding='utf-8') as f:
-            translated_text = f.read()
-            # Split by double newlines if segments are separated that way, 
-            # or adjust based on how segments were saved
-            validation_job.translated_segments = translated_text.split('\n')
-            # Filter out empty strings from the list
-            validation_job.translated_segments = [s for s in validation_job.translated_segments if s.strip()]
-        
-        # Ensure we have matching segment counts
-        if len(validation_job.segments) != len(validation_job.translated_segments):
-            print(f"Warning: Segment count mismatch - Source: {len(validation_job.segments)}, Translated: {len(validation_job.translated_segments)}")
-            # Try to match segments by combining translated lines if needed
-            if len(validation_job.translated_segments) > len(validation_job.segments):
-                # Too many translated segments, might be due to line breaks
-                # Combine them to match source segment count
-                combined_segments = []
-                lines_per_segment = len(validation_job.translated_segments) // len(validation_job.segments)
-                for i in range(0, len(validation_job.translated_segments), lines_per_segment):
-                    combined_segments.append('\n'.join(validation_job.translated_segments[i:i+lines_per_segment]))
-                validation_job.translated_segments = combined_segments[:len(validation_job.segments)]
-        
-        # Also load the glossary from the job if available
-        if job.final_glossary:
-            validation_job.glossary = json.loads(job.final_glossary) if isinstance(job.final_glossary, str) else job.final_glossary
-        
-        # Define progress callback to update database
-        def update_progress(progress: int):
-            job.validation_progress = progress
-            db.commit()
-            print(f"--- [VALIDATION] Progress: {progress}% ---")
-        
-        # Reset progress at start
-        job.validation_progress = 0
-        db.commit()
-        
-        results, summary = validator.validate_job(
-            validation_job,
-            sample_rate=sample_rate,
-            quick_mode=quick_validation,
-            progress_callback=update_progress
-        )
-        
-        # The validate_job now returns a tuple, extract the report
-        report = {
-            'summary': summary,
-            'detailed_results': [r.to_dict() for r in results]
-        }
-        
-        # Save validation report
-        os.makedirs("validation_logs", exist_ok=True)
-        report_filename = f"{os.path.splitext(job.filename)[0]}_validation_report.json"
-        report_path = os.path.join("validation_logs", report_filename)
-        
-        with open(report_path, 'w', encoding='utf-8') as f:
-            json.dump(report, f, ensure_ascii=False, indent=2)
-        
-        # Update job with results
-        job.validation_status = "COMPLETED"
-        job.validation_progress = 100
-        job.validation_report_path = report_path
-        job.validation_completed_at = func.now()
-        db.commit()
-        
-        print(f"--- [VALIDATION] Completed validation for Job ID: {job_id} ---")
-        
-    except Exception as e:
-        if db:
-            job = crud.get_job(db, job_id=job_id)
-            if job:
-                job.validation_status = "FAILED"
-                db.commit()
-        print(f"--- [VALIDATION] Error validating Job ID {job_id}: {e} ---")
-        traceback.print_exc()
-    finally:
-        if db:
-            db.close()
-        gc.collect()
-
-def run_post_edit_in_background(job_id: int, file_path: str, validation_report_path: str):
-    """Background task to run post-editing on a validated translation."""
-    from core.translation.post_editor import PostEditEngine
-    
-    db = None
-    try:
-        db = SessionLocal()
-        job = crud.get_job(db, job_id=job_id)
-        if not job:
-            print(f"--- [POST-EDIT] Job ID {job_id} not found ---")
-            return
-        
-        # Update status to IN_PROGRESS
-        job.post_edit_status = "IN_PROGRESS"
-        db.commit()
-        print(f"--- [POST-EDIT] Starting post-editing for Job ID: {job_id} ---")
-        
-        # Get the translated file path
-        unique_base = os.path.splitext(os.path.basename(file_path))[0]
-        original_ext = os.path.splitext(job.filename)[1].lower()
-        
-        if original_ext == '.epub':
-            output_ext = '.epub'
-        else:
-            output_ext = '.txt'
-        
-        translated_filename = f"{unique_base}_translated{output_ext}"
-        translated_path = os.path.join("translated_novel", translated_filename)
-        
-        if not os.path.exists(translated_path):
-            raise FileNotFoundError(f"Translated file not found: {translated_path}")
-        
-        # Initialize post-editor
-        config = load_config()
-        # Get API key from the job's owner (would need to store this or pass it)
-        # For now, we'll use the environment variable as fallback
-        api_key = os.environ.get("GEMINI_API_KEY", "")
-        model_name = "gemini-2.0-flash-exp"
-        
-        model_api = get_model_api(api_key, model_name, config)
-        
-        # Get glossary if available
-        glossary = job.final_glossary if job.final_glossary else {}
-        
-        post_editor = PostEditEngine(model_api, glossary=glossary)
-        
-        # Run post-editing
-        translation_job = TranslationJob(file_path, original_filename=job.filename)
-        
-        postedited_path = post_editor.post_edit_job(
-            translation_job,
-            translated_path,
-            validation_report_path
-        )
-        
-        # Get the post-edit log path
-        log_filename = f"{os.path.splitext(job.filename)[0]}_postedit_log.json"
-        log_path = os.path.join("postedit_logs", log_filename)
-        
-        # Update job with results
-        job.post_edit_status = "COMPLETED"
-        job.post_edit_log_path = log_path
-        job.post_edit_completed_at = func.now()
-        db.commit()
-        
-        print(f"--- [POST-EDIT] Completed post-editing for Job ID: {job_id} ---")
-        
-    except Exception as e:
-        if db:
-            job = crud.get_job(db, job_id=job_id)
-            if job:
-                job.post_edit_status = "FAILED"
-                db.commit()
-        print(f"--- [POST-EDIT] Error post-editing Job ID {job_id}: {e} ---")
-        traceback.print_exc()
-    finally:
-        if db:
-            db.close()
-        gc.collect()
-
-# --- API Endpoints ---
-
+# Root endpoint
 @app.get("/")
 def read_root():
-    return {"message": "Translation Service Backend is running!"}
-
-# --- Translation Endpoints ---
-
-class StyleAnalysisResponse(BaseModel):
-    protagonist_name: str
-    narration_style_endings: str
-    tone_keywords: str
-    stylistic_rule: str
-
-class GlossaryTerm(BaseModel):
-    term: str
-    translation: str
-
-class GlossaryAnalysisResponse(BaseModel):
-    glossary: List[GlossaryTerm]
-
-@app.post("/api/v1/analyze-style", response_model=StyleAnalysisResponse)
-async def analyze_style(
-    file: UploadFile = File(...),
-    api_key: str = Form(...),
-    model_name: str = Form("gemini-2.5-flash-lite"),
-    # No longer requires user authentication for this specific endpoint
-):
-    if not validate_api_key(api_key, model_name):
-        raise HTTPException(status_code=400, detail="Invalid API Key or unsupported model.")
-
-    temp_dir = "uploads"
-    os.makedirs(temp_dir, exist_ok=True)
-    # Create a unique temporary filename to avoid collisions
-    unique_id = uuid.uuid4()
-    temp_file_path = os.path.join(temp_dir, f"temp_{unique_id}_{file.filename}")
-    try:
-        with open(temp_file_path, "wb") as buffer:
-            shutil.copyfileobj(file.file, buffer)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to save temporary file: {e}")
-
-    try:
-        # Use centralized function to extract sample text (matching core engine approach)
-        initial_text = extract_sample_text(temp_file_path, method="first_segment", count=15000)
-
-        config = load_config()
-        model_api = get_model_api(api_key, model_name, config)
-        
-        print("\n--- Defining Core Narrative Style via API... ---")
-        # Use centralized function to analyze style
-        style_report_text = analyze_narrative_style_with_api(initial_text, model_api, file.filename)
-        print(f"Style defined as: {style_report_text}")
-
-        # Use centralized function to parse style analysis
-        parsed_style = parse_style_analysis(style_report_text)
-        
-        if len(parsed_style) < 3:
-            found_keys = list(parsed_style.keys())
-            error_detail = (
-                "Failed to parse all style attributes from the report. "
-                f"Successfully parsed: {found_keys}. "
-                f"Received from AI: '{style_report_text}'"
-            )
-            raise HTTPException(status_code=500, detail=error_detail)
-
-        return StyleAnalysisResponse(**parsed_style)
-    except Exception as e:
-        traceback.print_exc()
-        # Pass the specific error message to the frontend
-        raise HTTPException(status_code=500, detail=str(e))
-    finally:
-        if os.path.exists(temp_file_path):
-            os.remove(temp_file_path)
-
-
-@app.post("/api/v1/analyze-glossary", response_model=GlossaryAnalysisResponse)
-async def analyze_glossary(
-    file: UploadFile = File(...),
-    api_key: str = Form(...),
-    model_name: str = Form("gemini-2.5-flash-lite"),
-):
-    if not validate_api_key(api_key, model_name):
-        raise HTTPException(status_code=400, detail="Invalid API Key or unsupported model.")
-
-    temp_dir = "uploads"
-    os.makedirs(temp_dir, exist_ok=True)
-    unique_id = uuid.uuid4()
-    temp_file_path = os.path.join(temp_dir, f"temp_{unique_id}_{file.filename}")
-    try:
-        with open(temp_file_path, "wb") as buffer:
-            shutil.copyfileobj(file.file, buffer)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to save temporary file: {e}")
-
-    try:
-        initial_text = extract_sample_text(temp_file_path, method="first_segment", count=15000)
-        config = load_config()
-        model_api = get_model_api(api_key, model_name, config)
-        
-        print("\n--- Extracting Glossary via API... ---")
-        glossary_report_text = analyze_glossary_with_api(initial_text, model_api, file.filename)
-        print(f"Glossary extracted as: {glossary_report_text}")
-
-        parsed_glossary = parse_glossary_analysis(glossary_report_text)
-        
-        if not parsed_glossary:
-            # It's okay to return an empty list if no terms were found
-            return GlossaryAnalysisResponse(glossary=[])
-
-        return GlossaryAnalysisResponse(glossary=parsed_glossary)
-    except Exception as e:
-        traceback.print_exc()
-        # Pass the specific error message to the frontend
-        raise HTTPException(status_code=500, detail=str(e))
-    finally:
-        if os.path.exists(temp_file_path):
-            os.remove(temp_file_path)
-
-@app.post("/uploadfile/", response_model=schemas.TranslationJob)
-async def create_upload_file(
-    background_tasks: BackgroundTasks,
-    file: UploadFile = File(...),
-    api_key: str = Form(...),
-    model_name: str = Form("gemini-2.5-flash-lite"),
-    style_data: str = Form(None),
-    glossary_data: str = Form(None),
-    segment_size: int = Form(15000),
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    if not validate_api_key(api_key, model_name):
-        raise HTTPException(status_code=400, detail="Invalid API Key or unsupported model.")
-
-    job_create = schemas.TranslationJobCreate(filename=file.filename, owner_id=current_user.id)
-    db_job = crud.create_translation_job(db, job_create)
-    
-    sanitized_filename = re.sub(r'[^a-zA-Z0-9_.-]', '_', file.filename)
-    unique_filename = f"{db_job.id}_{sanitized_filename}"
-    file_path = f"uploads/{unique_filename}"
-    
-    os.makedirs("uploads", exist_ok=True)
-    try:
-        with open(file_path, "wb") as buffer:
-            shutil.copyfileobj(file.file, buffer)
-    except Exception as e:
-        crud.update_job_status(db, db_job.id, "FAILED", error_message=f"Failed to save file: {e}")
-        raise HTTPException(status_code=500, detail=f"Failed to save file: {e}")
-
-    crud.update_job_filepath(db, job_id=db_job.id, filepath=file_path)
-
-    background_tasks.add_task(run_translation_in_background, db_job.id, file_path, file.filename, api_key, model_name, style_data, glossary_data, segment_size)
-    
-    return db_job
-
-@app.get("/status/{job_id}", response_model=schemas.TranslationJob)
-def get_job_status(job_id: int, db: Session = Depends(get_db)):
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    return db_job
-
-@app.get("/download/{job_id}")
-async def download_translated_file(
-    job_id: int, 
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    
-    # Check ownership using Clerk User ID for consistency across environments
-    if not db_job.owner or db_job.owner.clerk_user_id != current_user.clerk_user_id:
-        # Also check for admin role as a fallback
-        user_is_admin = await auth.is_admin(current_user)
-        if not user_is_admin:
-            raise HTTPException(status_code=403, detail="Not authorized to download this file")
-
-    if db_job.status not in ["COMPLETED", "FAILED"]:
-        raise HTTPException(status_code=400, detail=f"Translation is not completed yet. Current status: {db_job.status}")
-    if not db_job.filepath:
-        raise HTTPException(status_code=404, detail="Filepath not found for this job.")
-
-    unique_base = os.path.splitext(os.path.basename(db_job.filepath))[0]
-    original_filename_base, original_ext = os.path.splitext(db_job.filename)
-
-    if original_ext.lower() == '.epub':
-        output_ext = '.epub'
-        media_type = 'application/epub+zip'
-    else:
-        output_ext = '.txt'
-        media_type = 'text/plain'
-
-    translated_unique_filename = f"{unique_base}_translated{output_ext}"
-    file_path = os.path.join("translated_novel", translated_unique_filename)
-
-    if not os.path.exists(file_path):
-        raise HTTPException(status_code=404, detail=f"Translated file not found at path: {file_path}")
-
-    user_translated_filename = f"{original_filename_base}_translated{output_ext}"
-
-    return FileResponse(path=file_path, filename=user_translated_filename, media_type=media_type)
-
-@app.get("/download/logs/{job_id}/{log_type}")
-async def download_log_file(
-    job_id: int, 
-    log_type: str, 
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-
-    # Check ownership using Clerk User ID for consistency across environments
-    if not db_job.owner or db_job.owner.clerk_user_id != current_user.clerk_user_id:
-        # Also check for admin role as a fallback
-        user_is_admin = await auth.is_admin(current_user)
-        if not user_is_admin:
-            raise HTTPException(status_code=403, detail="Not authorized to download logs for this file")
-
-    if log_type not in ["prompts", "context"]:
-        raise HTTPException(status_code=400, detail="Invalid log type. Must be 'prompts' or 'context'.")
-
-    base, _ = os.path.splitext(db_job.filename)
-    log_dir = "debug_prompts" if log_type == "prompts" else "context_log"
-    log_filename = f"{log_type}_job_{job_id}_{base}.txt"
-    log_path = os.path.join(log_dir, log_filename)
-
-    if not os.path.exists(log_path):
-        raise HTTPException(status_code=404, detail=f"{log_type.capitalize()} log file not found.")
-
-    return FileResponse(path=log_path, filename=log_filename, media_type="text/plain")
-
-@app.get("/api/v1/jobs/{job_id}/glossary")
-async def get_job_glossary(
-    job_id: int,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-
-    # Check ownership or admin role
-    user_is_admin = await auth.is_admin(current_user)
-    if not db_job.owner or (db_job.owner.clerk_user_id != current_user.clerk_user_id and not user_is_admin):
-        raise HTTPException(status_code=403, detail="Not authorized to access this glossary")
-
-    if db_job.status != "COMPLETED":
-        raise HTTPException(status_code=400, detail=f"Glossary is available only for completed jobs. Current status: {db_job.status}")
-
-    if not db_job.final_glossary:
-        return {} # Return an empty object if glossary is not set
-
-    return db_job.final_glossary
-
-# --- Validation and Post-Edit Endpoints ---
-
-@app.put("/api/v1/jobs/{job_id}/validation")
-async def trigger_validation(
-    job_id: int,
-    background_tasks: BackgroundTasks,
-    quick_validation: bool = Form(False),
-    validation_sample_rate: int = Form(100),  # percentage 0-100
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Trigger validation on a completed translation job."""
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    
-    # Check ownership or admin role
-    user_is_admin = await auth.is_admin(current_user)
-    if not db_job.owner or (db_job.owner.clerk_user_id != current_user.clerk_user_id and not user_is_admin):
-        raise HTTPException(status_code=403, detail="Not authorized to validate this job")
-    
-    if db_job.status != "COMPLETED":
-        raise HTTPException(status_code=400, detail=f"Can only validate completed jobs. Current status: {db_job.status}")
-    
-    # Update job with validation settings
-    db_job.validation_enabled = True
-    db_job.validation_status = "PENDING"
-    db_job.quick_validation = quick_validation
-    db_job.validation_sample_rate = validation_sample_rate
-    db.commit()
-    
-    # Add background task to run validation
-    background_tasks.add_task(run_validation_in_background, job_id, db_job.filepath, quick_validation, validation_sample_rate)
-    
-    return {"message": "Validation started", "job_id": job_id}
-
-@app.get("/api/v1/jobs/{job_id}/validation-report")
-async def get_validation_report(
-    job_id: int,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Get the validation report for a job."""
-    print(f"--- [API] Getting validation report for job {job_id} ---")
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    
-    # Check ownership or admin role
-    user_is_admin = await auth.is_admin(current_user)
-    if not db_job.owner or (db_job.owner.clerk_user_id != current_user.clerk_user_id and not user_is_admin):
-        raise HTTPException(status_code=403, detail="Not authorized to access this validation report")
-    
-    if db_job.validation_status != "COMPLETED":
-        print(f"--- [API] Validation not completed. Status: {db_job.validation_status} ---")
-        raise HTTPException(status_code=400, detail=f"Validation not completed. Current status: {db_job.validation_status}")
-    
-    print(f"--- [API] Validation report path: {db_job.validation_report_path} ---")
-    if not db_job.validation_report_path or not os.path.exists(db_job.validation_report_path):
-        print(f"--- [API] Report not found at path: {db_job.validation_report_path} ---")
-        raise HTTPException(status_code=404, detail="Validation report not found")
-    
-    # Read and return the JSON report
-    with open(db_job.validation_report_path, 'r', encoding='utf-8') as f:
-        report = json.load(f)
-    
-    print(f"--- [API] Successfully loaded validation report with {len(report.get('detailed_results', []))} results ---")
-    return report
-
-@app.put("/api/v1/jobs/{job_id}/post-edit")
-async def trigger_post_edit(
-    job_id: int,
-    background_tasks: BackgroundTasks,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Trigger post-editing on a validated translation job."""
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    
-    # Check ownership or admin role
-    user_is_admin = await auth.is_admin(current_user)
-    if not db_job.owner or (db_job.owner.clerk_user_id != current_user.clerk_user_id and not user_is_admin):
-        raise HTTPException(status_code=403, detail="Not authorized to post-edit this job")
-    
-    if db_job.validation_status != "COMPLETED":
-        raise HTTPException(status_code=400, detail="Validation must be completed before post-editing")
-    
-    if not db_job.validation_report_path:
-        raise HTTPException(status_code=400, detail="Validation report not found")
-    
-    # Update job with post-edit settings
-    db_job.post_edit_enabled = True
-    db_job.post_edit_status = "PENDING"
-    db.commit()
-    
-    # Add background task to run post-editing
-    background_tasks.add_task(run_post_edit_in_background, job_id, db_job.filepath, db_job.validation_report_path)
-    
-    return {"message": "Post-editing started", "job_id": job_id}
-
-@app.get("/api/v1/jobs/{job_id}/post-edit-log")
-async def get_post_edit_log(
-    job_id: int,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Get the post-edit log for a job."""
-    db_job = crud.get_job(db, job_id=job_id)
-    if db_job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-    
-    # Check ownership or admin role
-    user_is_admin = await auth.is_admin(current_user)
-    if not db_job.owner or (db_job.owner.clerk_user_id != current_user.clerk_user_id and not user_is_admin):
-        raise HTTPException(status_code=403, detail="Not authorized to access this post-edit log")
-    
-    if db_job.post_edit_status != "COMPLETED":
-        raise HTTPException(status_code=400, detail=f"Post-editing not completed. Current status: {db_job.post_edit_status}")
-    
-    if not db_job.post_edit_log_path or not os.path.exists(db_job.post_edit_log_path):
-        raise HTTPException(status_code=404, detail="Post-edit log not found")
-    
-    # Read and return the JSON log
-    with open(db_job.post_edit_log_path, 'r', encoding='utf-8') as f:
-        log = json.load(f)
-    
-    return log
-
-# --- Webhook Endpoints ---
-
-@app.post("/api/v1/webhooks/clerk")
-async def handle_clerk_webhook(request: Request, svix_id: str = Header(None), svix_timestamp: str = Header(None), svix_signature: str = Header(None), db: Session = Depends(get_db)):
-    if not CLERK_WEBHOOK_SECRET:
-        raise HTTPException(status_code=500, detail="Webhook secret is not configured.")
-    
-    headers = {
-        "svix-id": svix_id,
-        "svix-timestamp": svix_timestamp,
-        "svix-signature": svix_signature,
-    }
-    
-    try:
-        payload_body = await request.body()
-        wh = Webhook(CLERK_WEBHOOK_SECRET)
-        evt = wh.verify(payload_body, headers)
-    except Exception as e:
-        raise HTTPException(status_code=400, detail=f"Error verifying webhook signature: {e}")
-
-    event_type = evt["type"]
-    data = evt["data"]
-
-    if event_type == "user.created":
-        user_name = f'{data.get("first_name", "")} {data.get("last_name", "")}'.strip()
-        email_addresses = data.get("email_addresses", [])
-        email_address = email_addresses[0].get("email_address") if email_addresses else None
-
-        user_in = schemas.UserCreate(
-            clerk_user_id=data["id"],
-            email=email_address or None, # Ensure None if empty
-            name=user_name or None
-        )
-        crud.create_user(db, user=user_in)
-    elif event_type == "user.updated":
-        clerk_user_id = data["id"]
-        db_user = crud.get_user_by_clerk_id(db, clerk_id=clerk_user_id)
-        
-        user_name = f'{data.get("first_name", "")} {data.get("last_name", "")}'.strip()
-        email_addresses = data.get("email_addresses", [])
-        email_address = email_addresses[0].get("email_address") if email_addresses else None
-
-        if db_user:
-            user_update = schemas.UserUpdate(email=email_address or None, name=user_name or None)
-            crud.update_user(db, clerk_id=clerk_user_id, user_update=user_update)
-        else:
-            print(f"--- [INFO] Webhook received user.updated for non-existent user {clerk_user_id}. Creating them now. ---")
-            user_in = schemas.UserCreate(
-                clerk_user_id=clerk_user_id,
-                email=email_address or None, # Ensure None if empty
-                name=user_name or None
-            )
-            crud.create_user(db, user=user_in)
-    elif event_type == "user.deleted":
-        crud.delete_user(db, clerk_id=data["id"])
-        
-    return {"status": "success"}
-
-# --- SSE Announcement Stream Endpoint ---
-
-async def announcement_generator(request: Request):
-    last_sent_announcement = None
-    client_id = id(request)
-    print(f"📡 새 클라이언트 연결: {client_id}")
-
-    def get_announcement_from_db():
-        with SessionLocal() as db:
-            return crud.get_active_announcement(db)
-
-    try:
-        current_announcement = get_announcement_from_db()
-        if current_announcement:
-            announcement_data = {
-                "id": current_announcement.id,
-                "message": current_announcement.message,
-                "is_active": current_announcement.is_active,
-                "created_at": current_announcement.created_at.isoformat()
-            }
-            json_str = json.dumps(announcement_data, ensure_ascii=False)
-            yield f"data: {json_str}\n\n"
-            last_sent_announcement = current_announcement
-            print(f"📤 초기 공지 전송 (클라이언트 {client_id}): ID {current_announcement.id}")
-    except Exception as e:
-        print(f"❌ 초기 공지 전송 오류: {e}")
-
-    while True:
-        if await request.is_disconnected():
-            print(f"🔌 클라이언트 연결 해제: {client_id}")
-            break
-
-        try:
-            current_announcement = get_announcement_from_db()
-            
-            should_send = False
-            announcement_data = {}
-
-            if current_announcement is None and last_sent_announcement is not None:
-                announcement_data = {
-                    "id": last_sent_announcement.id,
-                    "message": last_sent_announcement.message,
-                    "is_active": False,
-                    "created_at": last_sent_announcement.created_at.isoformat()
-                }
-                should_send = True
-                last_sent_announcement = None
-                print(f"🔇 공지 비활성화 전송 (클라이언트 {client_id})")
-
-            elif current_announcement is not None:
-                if (last_sent_announcement is None or
-                    current_announcement.id != last_sent_announcement.id or
-                    current_announcement.message != last_sent_announcement.message or
-                    current_announcement.is_active != last_sent_announcement.is_active):
-                    
-                    announcement_data = {
-                        "id": current_announcement.id,
-                        "message": current_announcement.message,
-                        "is_active": current_announcement.is_active,
-                        "created_at": current_announcement.created_at.isoformat()
-                    }
-                    should_send = True
-                    last_sent_announcement = current_announcement
-                    print(f"📢 새 공지/변경 전송 (클라이언트 {client_id}): ID {current_announcement.id}")
-
-            if should_send:
-                json_str = json.dumps(announcement_data, ensure_ascii=False)
-                yield f"data: {json_str}\n\n"
-
-        except Exception as e:
-            print(f"❌ SSE 스트림 오류 (클라이언트 {client_id}): {e}")
-        
-        await asyncio.sleep(120)
-
-@app.get("/api/v1/announcements/stream")
-async def stream_announcements(request: Request):
-    return StreamingResponse(announcement_generator(request), media_type="text/event-stream; charset=utf-8")
-
-# --- Admin Endpoints ---
-
-def verify_admin_secret(x_admin_secret: str = Header(...)):
-    if x_admin_secret != ADMIN_SECRET_KEY:
-        raise HTTPException(status_code=403, detail="Invalid admin secret key")
-
-@app.post("/api/v1/admin/announcements", dependencies=[Depends(verify_admin_secret)])
-def create_new_announcement(announcement: schemas.AnnouncementCreate, db: Session = Depends(get_db)):
-    result = crud.create_announcement(db=db, announcement=announcement)
-    return JSONResponse(
-        content={
-            "id": result.id,
-            "message": result.message,
-            "is_active": result.is_active,
-            "created_at": result.created_at.isoformat()
-        },
-        media_type="application/json; charset=utf-8"
-    )
-
-@app.put("/api/v1/admin/announcements/{announcement_id}/deactivate", dependencies=[Depends(verify_admin_secret)])
-def deactivate_existing_announcement(announcement_id: int, db: Session = Depends(get_db)):
-    db_announcement = crud.deactivate_announcement(db, announcement_id)
-    if db_announcement is None:
-        raise HTTPException(status_code=404, detail="Announcement not found")
-    return JSONResponse(
-        content={
-            "id": db_announcement.id,
-            "message": db_announcement.message,
-            "is_active": db_announcement.is_active,
-            "created_at": db_announcement.created_at.isoformat()
-        },
-        media_type="application/json; charset=utf-8"
-    )
-
-@app.put("/api/v1/admin/announcements/deactivate-all", dependencies=[Depends(verify_admin_secret)])
-def deactivate_all_announcements(db: Session = Depends(get_db)):
-    """모든 활성 공지를 비활성화합니다."""
-    updated_count = crud.deactivate_all_announcements(db)
-    return JSONResponse(
-        content={
-            "message": f"모든 공지가 비활성화되었습니다.",
-            "deactivated_count": updated_count,
-            "success": True
-        },
-        media_type="application/json; charset=utf-8"
-    )
-
-
-# --- Community Board Endpoints ---
-
-# 임시 디버그 엔드포인트들은 보안상 제거됨
-
-@app.post("/api/v1/community/upload-image")
-async def upload_image(
-    file: UploadFile = File(...),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Upload an image for posts"""
-    # Validate file type
-    allowed_types = ["image/jpeg", "image/jpg", "image/png", "image/gif", "image/webp"]
-    if file.content_type not in allowed_types:
-        raise HTTPException(
-            status_code=400, 
-            detail="Only image files (JPEG, PNG, GIF, WebP) are allowed"
-        )
-    
-    # Validate file size (max 10MB)
-    file_content = await file.read()
-    if len(file_content) > 10 * 1024 * 1024:  # 10MB
-        raise HTTPException(status_code=400, detail="File size must be less than 10MB")
-    
-    # Generate unique filename
-    file_extension = os.path.splitext(file.filename)[1]
-    unique_filename = f"{uuid.uuid4()}{file_extension}"
-    file_path = os.path.join(UPLOAD_DIR, unique_filename)
-    
-    # Save file
-    with open(file_path, "wb") as buffer:
-        buffer.write(file_content)
-    
-    # Return the URL for accessing the uploaded image
-    image_url = f"/static/images/{unique_filename}"
-    
-    return {
-        "url": image_url,
-        "filename": unique_filename,
-        "original_name": file.filename,
-        "size": len(file_content)
-    }
-
-@app.get("/api/v1/community/categories", response_model=list[schemas.PostCategory])
-def get_categories(db: Session = Depends(get_db)):
-    """Get all post categories"""
-    return crud.get_post_categories(db)
-
-@app.get("/api/v1/community/categories/overview")
-def get_categories_with_recent_posts(
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_optional_user)
-):
-    """Get categories with their recent posts for community overview"""
-    categories = crud.get_post_categories(db)
-    categories_overview = []
-
-    # 모든 카테고리의 최근 게시물 3개를 한 번에 가져오기
-    all_recent_posts = []
-    for category in categories:
-        # Get 3 most recent posts for each category
-        recent_posts = crud.get_posts(db, category_id=category.id, skip=0, limit=3)
-        all_recent_posts.extend(recent_posts)
-
-    # 모든 최근 게시물의 댓글 수를 한 번의 쿼리로 가져오기
-    post_ids = [post.id for post in all_recent_posts]
-    comment_counts = crud.get_comment_counts_for_posts(db, post_ids)
-
-    for category in categories:
-        # 해당 카테고리의 최근 게시물 필터링
-        recent_posts = [post for post in all_recent_posts if post.category_id == category.id]
-        
-        # Convert posts to list format with proper schema
-        posts_data = []
-        for post in recent_posts:
-            can_view_private = crud.can_view_private_post(post, current_user)
-            
-            post_dict = {}
-            if can_view_private:
-                # User has permission, return full details
-                post_dict = {
-                    'id': post.id,
-                    'title': post.title,
-                    'author': post.author,
-                    'category': post.category,
-                    'is_pinned': post.is_pinned,
-                    'is_private': post.is_private,
-                    'view_count': post.view_count,
-                    'images': post.images or [],
-                    'created_at': post.created_at,
-                    'updated_at': post.updated_at,
-                    'comment_count': comment_counts.get(post.id, 0)
-                }
-            else:
-                # User does not have permission, mask the post
-                post_dict = {
-                    'id': post.id,
-                    'title': '🔒 비밀글입니다',
-                    'author': schemas.User(id=0, clerk_user_id="", name="익명", role="user", email="", created_at=post.created_at),
-                    'category': post.category,
-                    'is_pinned': post.is_pinned,
-                    'is_private': True,
-                    'view_count': post.view_count,
-                    'images': [],
-                    'created_at': post.created_at,
-                    'updated_at': post.updated_at,
-                    'comment_count': comment_counts.get(post.id, 0)
-                }
-            posts_data.append(post_dict)
-        
-        category_overview = {
-            'id': category.id,
-            'name': category.name,
-            'display_name': category.display_name,
-            'description': category.description,
-            'is_admin_only': category.is_admin_only,
-            'order': category.order,
-            'created_at': category.created_at,
-            'recent_posts': posts_data,
-            'total_posts': crud.count_posts(db, category_id=category.id)  # Use efficient count query
-        }
-        categories_overview.append(category_overview)
-    
-    return categories_overview
-
-@app.get("/api/v1/community/posts", response_model=list[schemas.PostList])
-def get_posts(
-    category: str = None,
-    skip: int = 0,
-    limit: int = 20,
-    search: str = None,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_optional_user)
-):
-    """Get posts, masking private ones based on user permissions."""
-    category_id = None
-    if category:
-        db_category = crud.get_post_category_by_name(db, category)
-        if db_category:
-            category_id = db_category.id
-    
-    posts = crud.get_posts(db, category_id=category_id, skip=skip, limit=limit, search=search)
-    
-    post_ids = [post.id for post in posts]
-    comment_counts = crud.get_comment_counts_for_posts(db, post_ids)
-    
-    sanitized_posts = []
-    for post in posts:
-        if crud.can_view_private_post(post, current_user):
-            post_data = schemas.PostList.model_validate(post)
-            post_data.comment_count = comment_counts.get(post.id, 0)
-            sanitized_posts.append(post_data)
-        else:
-            # Mask the private post
-            masked_post = schemas.PostList(
-                id=post.id,
-                title="🔒 비밀글입니다.",
-                author=schemas.User(id=0, clerk_user_id="", name="익명", role="user", email="", created_at=post.created_at),
-                category=post.category,
-                is_pinned=post.is_pinned,
-                is_private=True,
-                view_count=post.view_count,
-                images=[],
-                comment_count=comment_counts.get(post.id, 0),
-                created_at=post.created_at,
-                updated_at=post.updated_at
-            )
-            sanitized_posts.append(masked_post)
-            
-    return sanitized_posts
-
-@app.post("/api/v1/community/posts", response_model=schemas.Post)
-async def create_post(
-    post: schemas.PostCreate,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Create a new post"""
-    # Check if category is admin-only
-    category = db.query(models.PostCategory).filter(models.PostCategory.id == post.category_id).first()
-    if not category:
-        raise HTTPException(status_code=404, detail="Category not found")
-    
-    # 관리자 권한 확인 (로컬 DB + Clerk publicMetadata)
-    if category.is_admin_only:
-        user_is_admin = await auth.is_admin(current_user)
-        if not user_is_admin:
-            raise HTTPException(status_code=403, detail="Only admins can post in this category")
-    
-    return crud.create_post(db, post, current_user.id)
-
-@app.get("/api/v1/community/posts/{post_id}", response_model=schemas.Post)
-def get_post(
-    post_id: int, 
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_optional_user)
-):
-    """Get a specific post, masking private comments based on user permissions."""
-    post = crud.get_post(db, post_id)
-    if not post:
-        raise HTTPException(status_code=404, detail="Post not found")
-    
-    # Check if user can view the private post itself
-    if not crud.can_view_private_post(post, current_user):
-        # Unlike the list view, for a direct access, we raise an error
-        raise HTTPException(status_code=403, detail="Access denied to this private post")
-
-    # Sanitize comments within the post based on permissions
-    sanitized_comments = []
-    if post.comments:
-        for comment in post.comments:
-            if crud.can_view_private_comment(comment, current_user):
-                sanitized_comments.append(comment)
-            elif comment.is_private:
-                masked_comment = schemas.Comment(
-                    id=comment.id,
-                    content="🔒 비밀 댓글입니다.",
-                    author=schemas.User(id=0, clerk_user_id="", name="익명", role="user", email="", created_at=comment.created_at),
-                    post_id=comment.post_id,
-                    parent_id=comment.parent_id,
-                    is_private=True,
-                    created_at=comment.created_at,
-                    updated_at=comment.updated_at,
-                    replies=[]
-                )
-                sanitized_comments.append(masked_comment)
-            else:
-                sanitized_comments.append(comment)
-    
-    # Re-assign the sanitized comments to the post object before returning
-    post.comments = sanitized_comments
-    
-    return post
-
-@app.post("/api/v1/community/posts/{post_id}/view")
-def increment_post_view(
-    post_id: int, 
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_optional_user)
-):
-    """Increment post view count (separate endpoint)"""
-    post = crud.get_post(db, post_id)
-    if not post:
-        raise HTTPException(status_code=404, detail="Post not found")
-    
-    # Check if user can view private post
-    if not crud.can_view_private_post(post, current_user):
-        raise HTTPException(status_code=403, detail="Access denied to private post")
-    
-    updated_post = crud.increment_post_view_count(db, post_id)
-    return {"view_count": updated_post.view_count}
-
-@app.put("/api/v1/community/posts/{post_id}", response_model=schemas.Post)
-async def update_post(
-    post_id: int,
-    post_update: schemas.PostUpdate,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Update a post"""
-    post = crud.get_post(db, post_id)
-    if not post:
-        raise HTTPException(status_code=404, detail="Post not found")
-    
-    # Check permission (author or admin)
-    user_is_admin = await auth.is_admin(current_user)
-    if post.author_id != current_user.id and not user_is_admin:
-        raise HTTPException(status_code=403, detail="Not authorized to update this post")
-    
-    return crud.update_post(db, post_id, post_update)
-
-@app.delete("/api/v1/community/posts/{post_id}")
-async def delete_post(
-    post_id: int,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Delete a post"""
-    post = crud.get_post(db, post_id)
-    if not post:
-        raise HTTPException(status_code=404, detail="Post not found")
-    
-    # Check permission (author or admin)
-    user_is_admin = await auth.is_admin(current_user)
-    if post.author_id != current_user.id and not user_is_admin:
-        raise HTTPException(status_code=403, detail="Not authorized to delete this post")
-    
-    crud.delete_post(db, post_id)
-    return {"message": "Post deleted successfully"}
-
-# Comment endpoints
-@app.get("/api/v1/community/posts/{post_id}/comments", response_model=list[schemas.Comment])
-def get_comments(
-    post_id: int, 
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_optional_user)
-):
-    """Get comments for a post, masking private ones based on user permissions."""
-    post = crud.get_post(db, post_id)
-    if not post:
-        raise HTTPException(status_code=404, detail="Post not found")
-
-    # First, check if the user can view the post itself
-    if not crud.can_view_private_post(post, current_user):
-        raise HTTPException(status_code=403, detail="Access denied to this post and its comments")
-
-    comments = crud.get_comments(db, post_id)
-    
-    # Sanitize comments based on permissions
-    sanitized_comments = []
-    for comment in comments:
-        if crud.can_view_private_comment(comment, current_user):
-            # User has permission, return the full comment
-            sanitized_comments.append(comment)
-        elif comment.is_private:
-            # User cannot view, mask the content
-            masked_comment = schemas.Comment(
-                id=comment.id,
-                content="🔒 비밀 댓글입니다.",
-                author=schemas.User(id=0, clerk_user_id="", name="익명", role="user", email="", created_at=comment.created_at),
-                post_id=comment.post_id,
-                parent_id=comment.parent_id,
-                is_private=True,
-                created_at=comment.created_at,
-                updated_at=comment.updated_at,
-                replies=[] # Replies are handled recursively if needed
-            )
-            sanitized_comments.append(masked_comment)
-        else:
-            # It's a public comment, just add it
-            sanitized_comments.append(comment)
-            
-    return sanitized_comments
-
-@app.post("/api/v1/community/posts/{post_id}/comments", response_model=schemas.Comment)
-def create_comment(
-    post_id: int,
-    comment: schemas.CommentCreate,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Create a comment on a post"""
-    post = crud.get_post(db, post_id)
-    if not post:
-        raise HTTPException(status_code=404, detail="Post not found")
-    
-    # Ensure post_id matches
-    comment.post_id = post_id
-    
-    return crud.create_comment(db, comment, current_user.id)
-
-@app.put("/api/v1/community/comments/{comment_id}", response_model=schemas.Comment)
-async def update_comment(
-    comment_id: int,
-    comment_update: schemas.CommentUpdate,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Update a comment"""
-    comment = crud.get_comment(db, comment_id)
-    if not comment:
-        raise HTTPException(status_code=404, detail="Comment not found")
-    
-    # Check permission (author or admin)
-    user_is_admin = await auth.is_admin(current_user)
-    if comment.author_id != current_user.id and not user_is_admin:
-        raise HTTPException(status_code=403, detail="Not authorized to update this comment")
-    
-    return crud.update_comment(db, comment_id, comment_update)
-
-@app.delete("/api/v1/community/comments/{comment_id}")
-async def delete_comment(
-    comment_id: int,
-    db: Session = Depends(get_db),
-    current_user: models.User = Depends(auth.get_required_user)
-):
-    """Delete a comment"""
-    comment = crud.get_comment(db, comment_id)
-    if not comment:
-        raise HTTPException(status_code=404, detail="Comment not found")
-    
-    # Check permission (author or admin)
-    user_is_admin = await auth.is_admin(current_user)
-    if comment.author_id != current_user.id and not user_is_admin:
-        raise HTTPException(status_code=403, detail="Not authorized to delete this comment")
-    
-    crud.delete_comment(db, comment_id)
-    return {"message": "Comment deleted successfully"}
-
-# Admin endpoint to initialize categories
-@app.post("/api/v1/admin/community/init-categories", dependencies=[Depends(verify_admin_secret)])
-def initialize_categories(db: Session = Depends(get_db)):
-    """Initialize default post categories"""
-    default_categories = [
-        {"name": "notice", "display_name": "공지사항", "description": "중요한 공지사항", "is_admin_only": True, "order": 1},
-        {"name": "suggestion", "display_name": "건의사항", "description": "서비스 개선을 위한 제안", "is_admin_only": False, "order": 2},
-        {"name": "qna", "display_name": "Q&A", "description": "질문과 답변", "is_admin_only": False, "order": 3},
-        {"name": "free", "display_name": "자유게시판", "description": "자유로운 소통 공간", "is_admin_only": False, "order": 4}
-    ]
-    
-    created_categories = []
-    for cat_data in default_categories:
-        existing = crud.get_post_category_by_name(db, cat_data["name"])
-        if not existing:
-            category = crud.create_post_category(db, schemas.PostCategoryCreate(**cat_data))
-            created_categories.append(category)
-    
-    return {
-        "message": f"Created {len(created_categories)} categories",
-        "categories": created_categories
-    }
-
+    """Health check endpoint."""
+    return {"message": "Translation Service Backend is running!", "version": "1.0.0"}
