+++ backend/openapi.yaml
@@ -49,7 +49,7 @@ paths:
                     type: string
                     example: "Translation Service Backend is running!"
 
-  /api/v1/analyze-style:
+  /api/v1/translation/analyze-style:
     post:
       tags:
         - Translation
@@ -89,7 +89,7 @@ paths:
         '500':
           $ref: '#/components/responses/InternalServerError'
 
-  /api/v1/analyze-glossary:
+  /api/v1/translation/analyze-glossary:
     post:
       tags:
         - Translation
@@ -129,13 +129,13 @@ paths:
         '500':
           $ref: '#/components/responses/InternalServerError'
 
-  /uploadfile/:
+  /api/v1/translation/jobs:
     post:
       tags:
         - Translation
-      summary: Upload and translate file
-      description: Uploads a document file and starts the translation process
-      operationId: create_upload_file
+      summary: Create translation job
+      description: Creates a new translation job by uploading a document and starting the translation process
+      operationId: create_job
       security:
         - ClerkAuth: []
       requestBody:
@@ -183,13 +183,13 @@ paths:
         '500':
           $ref: '#/components/responses/InternalServerError'
 
-  /status/{job_id}:
+  /api/v1/translation/jobs/{job_id}:
     get:
       tags:
         - Translation
-      summary: Get translation job status
-      description: Retrieves the current status and progress of a translation job
-      operationId: get_job_status
+      summary: Get translation job
+      description: Retrieves the translation job details including status and progress
+      operationId: get_job
       parameters:
         - name: job_id
           in: path
@@ -207,13 +207,13 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /download/{job_id}:
+  /api/v1/translation/jobs/{job_id}/output:
     get:
       tags:
         - Translation
-      summary: Download translated file
-      description: Downloads the completed translation file
-      operationId: download_translated_file
+      summary: Download job output
+      description: Downloads the output file of a completed translation job
+      operationId: download_job_output
       security:
         - ClerkAuth: []
       parameters:
@@ -244,13 +244,13 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /download/logs/{job_id}/{log_type}:
+  /api/v1/translation/jobs/{job_id}/logs/{log_type}:
     get:
       tags:
         - Translation
-      summary: Download log files
+      summary: Download job logs
       description: Downloads debug log files for a translation job
-      operationId: download_log_file
+      operationId: download_job_log
       security:
         - ClerkAuth: []
       parameters:
@@ -284,7 +284,7 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /api/v1/jobs/{job_id}/glossary:
+  /api/v1/translation/jobs/{job_id}/glossary:
     get:
       tags:
         - Translation
@@ -318,13 +318,13 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /api/v1/jobs/{job_id}/validation:
+  /api/v1/translation/jobs/{job_id}/validation:
     put:
       tags:
         - Validation
       summary: Trigger validation
       description: Starts validation on a completed translation job
-      operationId: trigger_validation
+      operationId: validate_job
       security:
         - ClerkAuth: []
       parameters:
@@ -371,7 +371,7 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /api/v1/jobs/{job_id}/validation-report:
+  /api/v1/translation/jobs/{job_id}/validation-report:
     get:
       tags:
         - Validation
@@ -403,13 +403,13 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /api/v1/jobs/{job_id}/post-edit:
+  /api/v1/translation/jobs/{job_id}/post-edit:
     put:
       tags:
         - Validation
       summary: Trigger post-editing
       description: Starts post-editing on a validated translation job
-      operationId: trigger_post_edit
+      operationId: post_edit_job
       security:
         - ClerkAuth: []
       parameters:
@@ -440,7 +440,7 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /api/v1/jobs/{job_id}/post-edit-log:
+  /api/v1/translation/jobs/{job_id}/post-edit-log:
     get:
       tags:
         - Validation
@@ -523,7 +523,7 @@ paths:
         '500':
           $ref: '#/components/responses/InternalServerError'
 
-  /api/v1/announcements/stream:
+  /api/v1/announcements/sse:
     get:
       tags:
         - Admin
@@ -1000,7 +1000,7 @@ paths:
         '404':
           $ref: '#/components/responses/NotFound'
 
-  /api/v1/admin/community/init-categories:
+  /api/v1/admin/categories/initialize:
     post:
       tags:
         - Admin
